requirements.txt
flask
flask_sqlalchemy
flask_restful
flask-cors